Capability exceptions
Previously, we only handled Promise rejections arising from the read, locate,
and watch capabilities, but it is just as likely that a capability function
will throw. Now an exception is treated the same a rejection.

This change had the added benefit that the capabilities need not wrap their
return value in a Promise.