Txiki
This commit adds support for the Txiki (tjs) runtime.

Some code duplication was removed from the command-line REPLs. The "spec"
object was renamed "options", and the options are now grouped by environment.

Omitting the 'browser_port' option no longer causes a port to be allocated
automatically. Instead, the browser REPL is disabled. This change in behavior
seemed justified because it is painful to use the browser REPL with a dynamic
port, and there was no existing way to disable the browser REPL.