Parallelize digestion
Previously, hash generation was performed synchonously using the "node:crypto"
module, blocking the processing loop. This commit swaps out node:crypto for
crypto.subtle, which provides an eventual interface and presumably parallelizes
the work it is given using all available cores.

An exciting side effect of this change is that repl.js is now compatible with
the browser. Replete's core functionality could now be leveraged in a browser
environment, for example an browser-based IDE.

A downside is that the 'crypto' global only became available in Node.js v19, so
support for Node.js v18 has been dropped.

This commit also includes Replete's tests, now fit for public exhibition thanks
to 'import.meta.main'. To test a module, evaluate the whole file and check for
exceptions in the output.