--content_type
This commit adds the '--content_type' command line argument, making it possible
to easily and securely configure Replete as a Web development server.

As part of the change, 'options.mime' was replaced by 'options.headers' to
provide more control at the expense of increased verbosity. Backward
compatibility will be maintained for the time being.